Moved Embedder class to under neo4j_graphrag.embeddings
