+++ examples/hybrid_search.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from random import random
 
 from neo4j import GraphDatabase
-from neo4j_graphrag.embedder import Embedder
+from neo4j_graphrag.embeddings.embedder import Embedder
 from neo4j_graphrag.indexes import create_fulltext_index, create_vector_index
 from neo4j_graphrag.retrievers import HybridRetriever
