+++ src/neo4j_graphrag/embeddings/openai.py
@@ -17,7 +17,7 @@ from __future__ import annotations
 
 from typing import Any
 
-from neo4j_graphrag.embedder import Embedder
+from neo4j_graphrag.embeddings.embedder import Embedder
 
 
 class OpenAIEmbeddings(Embedder):
